+++ prov/efa/Makefile.include
@@ -51,7 +51,8 @@ _efa_files = \
 	prov/efa/src/rxr/rxr_msg.c	\
 	prov/efa/src/rxr/rxr_pkt_entry.c \
 	prov/efa/src/rxr/rxr_pkt_type_rts.c \
-	prov/efa/src/rxr/rxr_pkt_type_data.c
+	prov/efa/src/rxr/rxr_pkt_type_data.c \
+	prov/efa/src/rxr/rxr_pkt_type_misc.c
 
 _efa_headers = \
 	prov/efa/src/efa.h \
