+++ prov/efa/src/rxr/rxr_cq.c
@@ -436,45 +436,6 @@ write_err:
 	return 0;
 }
 
-void rxr_pkt_post_connack(struct rxr_ep *ep,
-			  struct rxr_peer *peer,
-			  fi_addr_t addr)
-{
-	struct rxr_pkt_entry *pkt_entry;
-	ssize_t ret;
-
-	if (peer->state == RXR_PEER_ACKED)
-		return;
-
-	pkt_entry = rxr_pkt_entry_alloc(ep, ep->tx_pkt_efa_pool);
-	if (OFI_UNLIKELY(!pkt_entry))
-		return;
-
-	rxr_ep_init_connack_pkt_entry(ep, pkt_entry, addr);
-
-	/*
-	 * TODO: Once we start using a core's selective completion capability,
-	 * post the CONNACK packets without FI_COMPLETION.
-	 */
-	ret = rxr_pkt_entry_send(ep, pkt_entry, addr);
-
-	/*
-	 * Skip sending this connack on error and try again when processing the
-	 * next RTS from this peer containing the source information
-	 */
-	if (OFI_UNLIKELY(ret)) {
-		rxr_pkt_entry_release_tx(ep, pkt_entry);
-		if (ret == -FI_EAGAIN)
-			return;
-		FI_WARN(&rxr_prov, FI_LOG_CQ,
-			"Failed to send a CONNACK packet: ret %zd\n", ret);
-	} else {
-		peer->state = RXR_PEER_ACKED;
-	}
-
-	return;
-}
-
 void rxr_cq_write_rx_completion(struct rxr_ep *ep,
 				struct rxr_rx_entry *rx_entry)
 {
@@ -742,94 +703,6 @@ void rxr_cq_handle_shm_rma_write_data(struct rxr_ep *ep, struct fi_cq_data_entry
 	rxr_release_rx_entry(ep, rx_entry);
 }
 
-/*
- * Sender handles the acknowledgment (RXR_EOR_PKT) from receiver on the completion
- * of the large message copy via fi_readmsg operation
- */
-static void rxr_cq_handle_eor(struct rxr_ep *ep,
-				  struct fi_cq_data_entry *comp,
-				  struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_eor_hdr *shm_eor;
-	struct rxr_tx_entry *tx_entry;
-
-	shm_eor = (struct rxr_eor_hdr *)pkt_entry->pkt;
-
-	/* pre-post buf used here, so can NOT track back to tx_entry with x_entry */
-	tx_entry = ofi_bufpool_get_ibuf(ep->tx_entry_pool, shm_eor->tx_id);
-	rxr_cq_write_tx_completion(ep, tx_entry);
-	rxr_pkt_entry_release_rx(ep, pkt_entry);
-}
-
-static void rxr_cq_handle_connack(struct rxr_ep *ep,
-				  struct fi_cq_data_entry *comp,
-				  struct rxr_pkt_entry *pkt_entry,
-				  fi_addr_t src_addr)
-{
-	struct rxr_peer *peer;
-
-	/*
-	 * We don't really need any information from the actual connack packet
-	 * itself, just the src_addr from the CQE
-	 */
-	assert(src_addr != FI_ADDR_NOTAVAIL);
-	peer = rxr_ep_get_peer(ep, src_addr);
-	peer->state = RXR_PEER_ACKED;
-	FI_DBG(&rxr_prov, FI_LOG_CQ,
-	       "CONNACK received from %" PRIu64 "\n", src_addr);
-	rxr_pkt_entry_release_rx(ep, pkt_entry);
-}
-
-static void rxr_cq_handle_readrsp(struct rxr_ep *ep,
-				  struct fi_cq_data_entry *comp,
-				  struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_readrsp_pkt *readrsp_pkt = NULL;
-	struct rxr_readrsp_hdr *readrsp_hdr = NULL;
-	struct rxr_rx_entry *rx_entry = NULL;
-
-	readrsp_pkt = (struct rxr_readrsp_pkt *)pkt_entry->pkt;
-	readrsp_hdr = &readrsp_pkt->hdr;
-	rx_entry = ofi_bufpool_get_ibuf(ep->rx_entry_pool, readrsp_hdr->rx_id);
-	assert(rx_entry->cq_entry.flags & FI_READ);
-	rx_entry->tx_id = readrsp_hdr->tx_id;
-	rxr_pkt_proc_data(ep, rx_entry, pkt_entry,
-			  readrsp_pkt->data,
-			  0, readrsp_hdr->seg_size);
-}
-
-static void rxr_cq_handle_cts(struct rxr_ep *ep,
-			      struct fi_cq_data_entry *comp,
-			      struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_peer *peer;
-	struct rxr_cts_hdr *cts_pkt;
-	struct rxr_tx_entry *tx_entry;
-
-	cts_pkt = (struct rxr_cts_hdr *)pkt_entry->pkt;
-	if (cts_pkt->flags & RXR_READ_REQ)
-		tx_entry = ofi_bufpool_get_ibuf(ep->readrsp_tx_entry_pool, cts_pkt->tx_id);
-	else
-		tx_entry = ofi_bufpool_get_ibuf(ep->tx_entry_pool, cts_pkt->tx_id);
-
-	tx_entry->rx_id = cts_pkt->rx_id;
-	tx_entry->window = cts_pkt->window;
-
-	/* Return any excess tx_credits that were borrowed for the request */
-	peer = rxr_ep_get_peer(ep, tx_entry->addr);
-	tx_entry->credit_allocated = ofi_div_ceil(cts_pkt->window, ep->max_data_payload_size);
-	if (tx_entry->credit_allocated < tx_entry->credit_request)
-		peer->tx_credits += tx_entry->credit_request - tx_entry->credit_allocated;
-
-	rxr_pkt_entry_release_rx(ep, pkt_entry);
-
-	if (tx_entry->state != RXR_TX_SEND) {
-		tx_entry->state = RXR_TX_SEND;
-		dlist_insert_tail(&tx_entry->entry, &ep->tx_pending_list);
-	}
-	return;
-}
-
 void rxr_cq_write_tx_completion(struct rxr_ep *ep,
 				struct rxr_tx_entry *tx_entry)
 {
@@ -965,19 +838,19 @@ void rxr_cq_handle_pkt_recv_completion(struct rxr_ep *ep,
 		rxr_pkt_handle_rts_recv(ep, pkt_entry);
 		return;
 	case RXR_EOR_PKT:
-		rxr_cq_handle_eor(ep, cq_entry, pkt_entry);
+		rxr_pkt_handle_eor_recv(ep, pkt_entry);
 		return;
 	case RXR_CONNACK_PKT:
-		rxr_cq_handle_connack(ep, cq_entry, pkt_entry, src_addr);
+		rxr_pkt_handle_connack_recv(ep, pkt_entry, src_addr);
 		return;
 	case RXR_CTS_PKT:
-		rxr_cq_handle_cts(ep, cq_entry, pkt_entry);
+		rxr_pkt_handle_cts_recv(ep, pkt_entry);
 		return;
 	case RXR_DATA_PKT:
 		rxr_pkt_handle_data_recv(ep, pkt_entry);
 		return;
 	case RXR_READRSP_PKT:
-		rxr_cq_handle_readrsp(ep, cq_entry, pkt_entry);
+		rxr_pkt_handle_readrsp_recv(ep, cq_entry, pkt_entry);
 		return;
 	default:
 		FI_WARN(&rxr_prov, FI_LOG_CQ,
@@ -1002,64 +875,6 @@ static int rxr_send_completion_mr_dereg(struct rxr_tx_entry *tx_entry)
 		}
 	}
 	return ret;
-}
-
-void rxr_cq_handle_rma_context_pkt(struct rxr_ep *ep, struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_tx_entry *tx_entry = NULL;
-	struct rxr_rx_entry *rx_entry = NULL;
-	struct rxr_rma_context_pkt *rma_context_pkt;
-	int ret;
-
-	assert(rxr_get_base_hdr(pkt_entry->pkt)->version == RXR_PROTOCOL_VERSION);
-
-	rma_context_pkt = (struct rxr_rma_context_pkt *)pkt_entry->pkt;
-
-	switch (rma_context_pkt->rma_context_type) {
-	case RXR_SHM_RMA_READ:
-	case RXR_SHM_RMA_WRITE:
-		/* Completion of RMA READ/WRTITE operations that apps call */
-		tx_entry = pkt_entry->x_entry;
-
-		if (tx_entry->fi_flags & FI_COMPLETION) {
-			rxr_cq_write_tx_completion(ep, tx_entry);
-		} else {
-			efa_cntr_report_tx_completion(&ep->util_ep, tx_entry->cq_entry.flags);
-			rxr_release_tx_entry(ep, tx_entry);
-		}
-		rxr_pkt_entry_release_tx(ep, pkt_entry);
-		break;
-	case RXR_SHM_LARGE_READ:
-		/*
-		 * This must be on the receiver (remote) side of two-sided message
-		 * transfer, which is also the initiator of RMA READ.
-		 * We get RMA READ completion for previously issued
-		 * fi_read operation over shm provider, which means
-		 * receiver side has received all data from sender
-		 */
-		rx_entry = pkt_entry->x_entry;
-		rx_entry->cq_entry.len = rx_entry->total_len;
-		rx_entry->bytes_done = rx_entry->total_len;
-
-		ret = rxr_ep_post_ctrl_or_queue(ep, RXR_TX_ENTRY, rx_entry, RXR_EOR_PKT, 1);
-		if (ret) {
-			if (rxr_cq_handle_rx_error(ep, rx_entry, ret))
-				assert(0 && "error writing error cq entry for EOR\n");
-		}
-
-		if (rx_entry->fi_flags & FI_MULTI_RECV)
-			rxr_msg_multi_recv_handle_completion(ep, rx_entry);
-		rxr_cq_write_rx_completion(ep, rx_entry);
-		rxr_msg_multi_recv_free_posted_entry(ep, rx_entry);
-		if (OFI_LIKELY(!ret))
-			rxr_release_rx_entry(ep, rx_entry);
-		rxr_pkt_entry_release_rx(ep, pkt_entry);
-		break;
-	default:
-		FI_WARN(&rxr_prov, FI_LOG_CQ, "invalid rma_context_type in RXR_RMA_CONTEXT_PKT %d\n",
-			rma_context_pkt->rma_context_type);
-		assert(0 && "invalid RXR_RMA_CONTEXT_PKT rma_context_type\n");
-	}
 }
 
 void rxr_cq_handle_pkt_send_completion(struct rxr_ep *ep, struct fi_cq_data_entry *comp)
@@ -1117,7 +932,7 @@ void rxr_cq_handle_pkt_send_completion(struct rxr_ep *ep, struct fi_cq_data_entr
 		tx_entry->bytes_acked += readrsp_hdr->seg_size;
 		break;
 	case RXR_RMA_CONTEXT_PKT:
-		rxr_cq_handle_rma_context_pkt(ep, pkt_entry);
+		rxr_pkt_handle_rma_context_send_completion(ep, pkt_entry);
 		return;
 	default:
 		FI_WARN(&rxr_prov, FI_LOG_CQ,
