+++ prov/efa/src/rxr/rxr_ep.c
@@ -612,104 +612,6 @@ void rxr_inline_mr_reg(struct rxr_domain *rxr_domain,
 	return;
 }
 
-void rxr_ep_calc_cts_window_credits(struct rxr_ep *ep, struct rxr_peer *peer,
-				    uint64_t size, int request,
-				    int *window, int *credits)
-{
-	struct efa_av *av;
-	int num_peers;
-
-	/*
-	 * Adjust the peer credit pool based on the current AV size, which could
-	 * have grown since the time this peer was initialized.
-	 */
-	av = rxr_ep_av(ep);
-	num_peers = av->used - 1;
-	if (num_peers && ofi_div_ceil(rxr_env.rx_window_size, num_peers) < peer->rx_credits)
-		peer->rx_credits = ofi_div_ceil(peer->rx_credits, num_peers);
-
-	/*
-	 * Allocate credits for this transfer based on the request, the number
-	 * of available data buffers, and the number of outstanding peers this
-	 * endpoint is actively tracking in the AV. Also ensure that a minimum
-	 * number of credits are allocated to the transfer so the sender can
-	 * make progress.
-	 */
-	*credits = MIN(MIN(ep->available_data_bufs, ep->posted_bufs_efa),
-		       peer->rx_credits);
-	*credits = MIN(request, *credits);
-	*credits = MAX(*credits, rxr_env.tx_min_credits);
-	*window = MIN(size, *credits * ep->max_data_payload_size);
-	if (peer->rx_credits > ofi_div_ceil(*window, ep->max_data_payload_size))
-		peer->rx_credits -= ofi_div_ceil(*window, ep->max_data_payload_size);
-}
-
-ssize_t rxr_pkt_init_cts(struct rxr_ep *ep,
-			 struct rxr_rx_entry *rx_entry,
-			 struct rxr_pkt_entry *pkt_entry)
-{
-	int window = 0;
-	struct rxr_cts_hdr *cts_hdr;
-	struct rxr_peer *peer;
-	size_t bytes_left;
-
-	cts_hdr = (struct rxr_cts_hdr *)pkt_entry->pkt;
-	cts_hdr->type = RXR_CTS_PKT;
-	cts_hdr->version = RXR_PROTOCOL_VERSION;
-	cts_hdr->flags = 0;
-
-	if (rx_entry->cq_entry.flags & FI_READ)
-		cts_hdr->flags |= RXR_READ_REQ;
-
-	cts_hdr->tx_id = rx_entry->tx_id;
-	cts_hdr->rx_id = rx_entry->rx_id;
-
-	bytes_left = rx_entry->total_len - rx_entry->bytes_done;
-	peer = rxr_ep_get_peer(ep, rx_entry->addr);
-	rxr_ep_calc_cts_window_credits(ep, peer, bytes_left,
-				       rx_entry->credit_request,
-				       &window, &rx_entry->credit_cts);
-	cts_hdr->window = window;
-	pkt_entry->pkt_size = RXR_CTS_HDR_SIZE;
-	pkt_entry->addr = rx_entry->addr;
-	pkt_entry->x_entry = (void *)rx_entry;
-	return 0;
-}
-
-void rxr_pkt_handle_cts_sent(struct rxr_ep *ep,
-			     struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_rx_entry *rx_entry;
-
-	rx_entry = (struct rxr_rx_entry *)pkt_entry->x_entry;
-	rx_entry->window = rxr_get_cts_hdr(pkt_entry->pkt)->window;
-	ep->available_data_bufs -= rx_entry->credit_cts;
-
-	/*
-	 * Set a timer if available_bufs is exhausted. We may encounter a
-	 * scenario where a peer has stopped responding so we need a fallback
-	 * to replenish the credits.
-	 */
-	if (OFI_UNLIKELY(ep->available_data_bufs == 0))
-		ep->available_data_bufs_ts = ofi_gettime_us();
-}
-
-void rxr_ep_init_connack_pkt_entry(struct rxr_ep *ep,
-				   struct rxr_pkt_entry *pkt_entry,
-				   fi_addr_t addr)
-{
-	struct rxr_connack_hdr *connack_hdr;
-
-	connack_hdr = (struct rxr_connack_hdr *)pkt_entry->pkt;
-
-	connack_hdr->type = RXR_CONNACK_PKT;
-	connack_hdr->version = RXR_PROTOCOL_VERSION;
-	connack_hdr->flags = 0;
-
-	pkt_entry->pkt_size = RXR_CONNACK_HDR_SIZE;
-	pkt_entry->addr = addr;
-}
-
 int rxr_ep_init_ctrl_pkt(struct rxr_ep *rxr_ep, int entry_type, void *x_entry,
 			 int ctrl_type, struct rxr_pkt_entry *pkt_entry)
 {
