+++ prov/efa/src/rxr/rxr_pkt_type.h
@@ -180,7 +180,6 @@ void rxr_pkt_post_connack(struct rxr_ep *ep,
 			  fi_addr_t addr);
 
 void rxr_pkt_handle_connack_recv(struct rxr_ep *ep,
-				 struct fi_cq_data_entry *comp,
 				 struct rxr_pkt_entry *pkt_entry,
 				 fi_addr_t addr);
 /*
@@ -211,6 +210,10 @@ struct rxr_cts_hdr *rxr_get_cts_hdr(void *pkt)
 	return (struct rxr_cts_hdr *)pkt;
 }
 
+void rxr_pkt_calc_cts_window_credits(struct rxr_ep *ep, struct rxr_peer *peer,
+				     uint64_t size, int request,
+				     int *window, int *credits);
+
 ssize_t rxr_pkt_init_cts(struct rxr_ep *ep,
 			 struct rxr_rx_entry *rx_entry,
 			 struct rxr_pkt_entry *pkt_entry);
@@ -219,7 +222,6 @@ void rxr_pkt_handle_cts_sent(struct rxr_ep *ep,
 			     struct rxr_pkt_entry *pkt_entry);
 
 void rxr_pkt_handle_cts_recv(struct rxr_ep *ep,
-			     struct fi_cq_data_entry *comp,
 			     struct rxr_pkt_entry *pkt_entry);
 
 /*
